Consolidate shared contact and config loaders
