+++ agrimate-website/server/config.php
@@ -1,4 +1,5 @@
 <?php
-require_once __DIR__ . '/config_loader.php';
+
+require_once dirname(__DIR__, 2) . '/server/shared/config_loader.php';
 
 return agrimate_load_config();
