+++ agrimate-website/server/contact.php
@@ -1,5 +1,5 @@
 <?php
 
-require_once __DIR__ . '/contact_handler.php';
+require_once dirname(__DIR__, 2) . '/server/shared/contact_handler.php';
 
 agrimate_handle_contact_request();
